Tweak 2nd layer for more consistent mods
Welp... I knew I'd want these in their proper places on the second layer
since I use it for several shorcuts and such. I moved some of the
special punctionation characters to the right for keys that aren't used
in combinations and are easily accessible on the base layer, which is
just a key release away.