Even close to the atreus setup
This essentially takes the outer three keys on each side of the bottom
row from the atreus and arranges them vertically in the extra column
that the zen has. I was most of the way there already, but this gets us
all the way there, which should make all of the little interactions just
a bit easier.

So far, the mod-tap functionality on the zen has been much better than
the atreus, so I also have shift back on both sides via mod-tap.