Put escape back on the thumb
I've already gotten used to it enough that I think it'll be a good move.
Tab is still in its normal spot and escape is still on the bottom row
pinky, but that's feeling more and more awkward.  I also like the
symmetry between esc and return on the thumbs.