Give tab its own key for use as modifier
Tab was fine as a tap on an mt key, but now I have need to use it as a
modifier (thanks Asana). This little rearrangement should be just fine
as it mostly moves alt to its own key. It also has the effect of moving
F13 (for Shush) to a key that can also be held, which is handy.